requirements.txt
# PDF processing
PyPDF2>=3.0.0
PyMuPDF>=1.23.0

# OCR support
pytesseract>=0.3.10
pdf2image>=3.1.0
Pillow>=9.0.0

# Text processing
nltk>=3.8
spacy>=3.4.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Machine learning
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# Database
#sqlite3  # Built-in

# Utilities
tqdm>=4.64.0
psutil>=5.9.0  # Memory monitoring
pathlib  # Built-in
dataclasses  # Built-in
